[canvaskit] make picture disposal work on older browsers (#26981)
* [canvaskit] make picture disposal work on browsers that don't yet support FinalizationRegistry

+++ lib/web_ui/lib/src/engine/canvaskit/picture.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:ui/src/engine.dart';
 import 'package:ui/ui.dart' as ui;
 
 import 'canvas.dart';
@@ -9,6 +10,11 @@ import 'canvaskit_api.dart';
 import 'image.dart';
 import 'skia_object_cache.dart';
 
+/// Implements [ui.Picture] on top of [SkPicture].
+///
+/// Unlike most other [ManagedSkiaObject] implementations, instances of this
+/// class may have their Skia counterparts deleted before finalization registry
+/// or [SkiaObjectCache] decide to delete it.
 class CkPicture extends ManagedSkiaObject<SkPicture> implements ui.Picture {
   final ui.Rect? cullRect;
   final CkPictureSnapshot? _snapshot;
@@ -24,14 +30,36 @@ class CkPicture extends ManagedSkiaObject<SkPicture> implements ui.Picture {
   @override
   int get approximateBytesUsed => 0;
 
+  /// Whether the picture has been disposed of.
+  ///
+  /// This is indended to be used in tests and assertions only.
+  bool get debugIsDisposed => _isDisposed;
+
+  /// This is set to true when [dispose] is called and is never reset back to
+  /// false.
+  ///
+  /// This extra flag is necessary on top of [rawSkiaObject] because
+  /// [rawSkiaObject] being null does not indicate permanent deletion. See
+  /// similar flag [SkiaObjectBox.isDeletedPermanently].
+  bool _isDisposed = false;
+
   @override
   void dispose() {
+    assert(!_isDisposed, 'Object has been disposed.');
+    if (Instrumentation.enabled) {
+      Instrumentation.instance.incrementCounter('Picture disposed');
+    }
+    _isDisposed = true;
     _snapshot?.dispose();
-    skiaObject.delete();
+
+    // Emulate what SkiaObjectCache does.
+    rawSkiaObject?.delete();
+    rawSkiaObject = null;
   }
 
   @override
   Future<ui.Image> toImage(int width, int height) async {
+    assert(!_isDisposed);
     final SkSurface skSurface = canvasKit.MakeSurface(width, height);
     final SkCanvas skCanvas = skSurface.getCanvas();
     skCanvas.drawPicture(skiaObject);
@@ -51,11 +79,19 @@ class CkPicture extends ManagedSkiaObject<SkPicture> implements ui.Picture {
 
   @override
   SkPicture resurrect() {
+    // If a picture has been explicitly disposed of, it can no longer be
+    // resurrected. An attempt to resurrect after the framework told the
+    // engine to dispose of the picture likely indicates a bug in the engine.
+    assert(!_isDisposed);
     return _snapshot!.toPicture();
   }
 
   @override
   void delete() {
-    rawSkiaObject?.delete();
+    // This method may be called after [dispose], in which case there's nothing
+    // left to do. The Skia object is deleted permanently.
+    if (!_isDisposed) {
+      rawSkiaObject?.delete();
+    }
   }
 }
